cleaner/requirements.txt
loguru==0.7.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3